switched user, item, and centroid weights to doubles

+++ cpp/algo.hpp
@@ -16,13 +16,13 @@ int find_theta_bin_index(const float theta_uc,
                          const std::vector<float> theta_bins,
                          const int num_bins);
 void computeTopKForCluster(
-    const int cluster_id, const float *centroid,
-    const std::vector<int> &user_ids_in_cluster, const float *user_weights,
-    const float *item_weights, const float *item_norms, const float *theta_ics,
+    const int cluster_id, const double *centroid,
+    const std::vector<int> &user_ids_in_cluster, const double *user_weights,
+    const double *item_weights, const float *item_norms, const float *theta_ics,
     const float &centroid_norm, const int num_items,
     const int num_latent_factors, const int num_bins, const int K,
     const int batch_size, float *upper_bounds, int *sorted_upper_bounds_indices,
-    float *sorted_upper_bounds, float *sorted_item_weights,
+    float *sorted_upper_bounds, double *sorted_item_weights,
     std::ofstream &user_stats_file);
 
 #endif /* algo_hpp */
