+++ cpp/parser.hpp
@@ -11,8 +11,8 @@
 
 #include <string>
 
-float *parse_weights_csv(const std::string filename, const int num_rows,
-                         const int num_cols);
+double *parse_weights_csv(const std::string filename, const int num_rows,
+                          const int num_cols);
 int *parse_ids_csv(const std::string filename, const int num_rows);
 
 #endif /* parser_hpp */
